requirements.txt
fastapi uvicorn 
python-dotenv
httpx
pydantic
langgraph
langchain
google-generativeai
pycountry
ffmpeg-python
faster-whisper
apscheduler